requirements.txt
asgiref==3.11.0
Django==5.2.8
django-hashid-field==3.4.1
djangorestframework==3.16.1
hashids==1.3.1
sqlparse==0.5.4
